feat(research+ui) gpt5mini fixed - not great but working
feat(research+ui): add MeSH/synonym expansion + multi-query PubMed, CT.gov
  fallbacks; preserve per-source debug & structured research_papers_data;
  restore inline lit-search UI and stable References table view.
