+++ api.py
@@ -1,0 +1,4 @@
+# Top-level shim to import FastAPI app and helpers from src package for tests/imports
+from src.api import app, process_idea, analyze_scenarios, perform_statistical_calculations, map_study_type_to_test
+
+__all__ = ["app", "process_idea", "analyze_scenarios", "perform_statistical_calculations", "map_study_type_to_test"]
